Added sorrygoogle.userscript.js
Whenever I'm on a VPN, Google tends to throw me to their Captcha page (the one with /sorry in the URL). They made these captchas quite hard, to the point where it sometimes would take me minutes to solve these. DuckDuckGo does not give you a captcha to solve when using a VPN, so I usually go there, but then I have tp type the query out again. This userscript extracts your query from the URL and offers a link to DDG with the query prefilled. Suck it, Google.