Create nitwit.userscript.js
Injects a button on the top right to quickly toggle between Twitter (apparently now known as X, though I fail to see why) and Nitter (or any other instance). If you use any other instance, be sure to update or add @match as well. For example, I like unofficialbird.com, which is what's in there.

I'm fairly new at this so if it doesn't work for you or if you have any suggestions, please leave some feedback and I'll take a look when I get a chance.